\name{Data_Erodibility}
 
\alias{Data_Erodibility}

\title{Erodibility dataset.}

\description{Dataset of physical and chemical attributes used in estimating soil erodibility.}

\usage{data(Data_Erodibility)}

\format{Physical and chemical attributes used in estimating soil erodibility.}

\references{
Denardin, J. E. (1990). “Erodibilidade de solo estimada por meio de parâmetros físicos e químicos”. Piracicaba, ESALQ, 1990. 81p. (Tese de Doutorado). <https:...>

RadamBrasil, P. (1981a). Folha SC. 22 Tocantins, Rio de Janeiro: Ministério das Minas e Energia. 524p. <https:...>

RadamBrasil, P. (1981b). FOLHA SD. 22 Goiás, Rio de Janeiro: Ministério das Minas e Energia. 640p. <https:...>

RadamBrasil, P. (1982). Folha SD. 23: Brasília, Rio de Janeiro: Ministério das Minas e Energia. 660p. <https:...>

RadamBrasil, P. (1983). Folha SE. 22 Goiânia, Rio de Janeiro: Ministério das Minas e Energia. 768p. <https:...>
}

\author{
Dione Pereira Cardoso\cr

Paulo Cesar Ossani\cr

Junior Cesar Avanzi
}

\examples{
data(Data_Erodibility)
head(Data_Erodibility)
}

\keyword{Data set}
