\name{Data_Rainfall_minutes}
 
\alias{Data_Rainfall_minutes}

\title{Rainfall dataset.}

\description{Sub-hourly rainfall data for the municipality of Peixe, TO, for the period January to December 2023.}

\usage{data(Data_Rainfall_minutes)}

\format{Data set with 22,032 observations with 3 variables, referring to precipitation from January to December 2023, in the municipality of Peixe, TO, Brazil. The columns being: date, times and rainfall.}

\references{
Cemaden (2024). Centro nacional de monitoramento e alertas de desastres naturais - <http://www.cemaden.gov.br/apresentacao/>.
}

\author{
Dione Pereira Cardoso\cr

Paulo Cesar Ossani\cr

Junior Cesar Avanzi
}

\examples{
data(Data_Rainfall_minutes)
head(Data_Rainfall_minutes)
}

\keyword{Data set}
