\name{Data_Rainfall_month}
 
\alias{Data_Rainfall_month}

\title{Rainfall dataset.}

\description{Monthly rainfall data for the municipality of Peixe, TO, for the period from 2013 to 2023 (Source: BDMEP-INMET, 2024).}

\usage{data(Data_Rainfall_month)}

\format{Dataset monthly referring to rainfall in 2013 and 2023, in the municipality of Peixe, TO, Brazil.}

\references{
INMET - Instituto Nacional de Meteorologia (2018). “BDMEP - Banco de Dados Meteorológicos para Ensino e Pesquisa - Série Histórica - Dados Mensais – Precipitação (mm)”. Brasília.
}

\author{
Dione Pereira Cardoso\cr

Paulo Cesar Ossani\cr

Junior Cesar Avanzi
}

\examples{
data(Data_Rainfall_month)
head(Data_Rainfall_month)
}

\keyword{Data set}
