\name{Data_SoilLoss}
 
\alias{Data_SoilLoss}

\title{Water erosion dataset.}

\description{Dataset of erosivity, erodibility, topography, LULC, and support conservation practices.}

\usage{data(Data_SoilLoss)}

\format{Dataset of erosivity, erodibility, topography, LULC, and support conservation practices of several years.}

\author{
Dione Pereira Cardoso\cr

Paulo Cesar Ossani\cr

Junior Cesar Avanzi
}

\examples{
data(Data_SoilLoss)
head(Data_SoilLoss)
}

\keyword{Data set}
