\name{SoilConservation-package}

\alias{SoilConservation-package}

\docType{package}

\title{Soil and Water Conservation.}

\description{Includes four functions: RFactor_calc(), RFactor_est(), KFactor() and SoilLoss(). The rainfall erosivity factors can be calculated or estimated, and soil erodibility will be estimated by the equation extracted from the monograph. Soil loss will be estimated by the product of five factors (rainfall erosivity, soil erodibility, length and steepness slope, cover-management factor and support practice factor. In the future, additional functions can be included. This efforts to advance research in soil and water conservation, with fast and accurate results.}

\author{ 
 Dione Pereira Cardoso \email{cardoso.dione@gmail.com}\cr
 
 Paulo Cesar Ossani \email{ossanipc@hotmail.com}\cr
 
 Junior Cesar Avanzi \email{junior.avanzi@ufla.br}
}

\details{
 \tabular{ll}{
 Package: \tab SoilConservation\cr
 Type: \tab Package\cr
 Version: \tab 1.0.0\cr
 Date: \tab 2024-04-28\cr
 License: \tab GPL (>= 3)\cr\cr
 }
}

\references{
Arnoldus H. M. J. (1980). An approximation of the rainfall factor in the universal soil loss equation. In: De Boodt M, Gabriels D (eds) Assessment of erosion. JohnWiley & Sons, Chichister, pp 127–132. <https:...>.

Brown, L.C. and Foster, G.R. (1987). Storm erosivity using idealized intensity distributions. Trans. ASAE 30, 2, 379–386. <https:...>.

Cemaden (2024). Centro nacional de monitoramento e alertas de desastres naturais - <http://www.cemaden.gov.br/apresentacao/>.

Denardin, J. E. (1990). “Erodibilidade de solo estimada por meio de parâmetros físicos e químicos”. Piracicaba, ESALQ, 1990. 81p. (Tese de Doutorado). <https:...>

Godoi, R. F.; Rodrigues, D. B.; Borrelli, P.; Oliveira, P. T. S. (2021). “High-resolution soil erodibility map of Brazil”. Science of The Total Environment, v. 781, p. 146673. <doi:10.1016/j.scitotenv.2021.146673>.

INMET - Instituto Nacional de Meteorologia (2018). “BDMEP - Banco de Dados Meteorológicos para Ensino e Pesquisa - Série Histórica - Dados Mensais – Precipitação (mm)”. Brasília.

Oliveira, P. T. S. and Wendland, E.; Nearing, M. A. (2013). “Rainfall erosivity in Brazil: A review”. Catena, v. 100, p. 139-147, 2013. <doi:10.1016/j.catena.2012.08.006>.

RadamBrasil, P. (1981a). Folha SC. 22 Tocantins, Rio de Janeiro: Ministério das Minas e Energia. 524p. <https:...>

RadamBrasil, P. (1981b). FOLHA SD. 22 Goiás, Rio de Janeiro: Ministério das Minas e Energia. 640p. <https:...>

RadamBrasil, P. (1982). Folha SD. 23: Brasília, Rio de Janeiro: Ministério das Minas e Energia. 660p. <https:...>

RadamBrasil, P. (1983). Folha SE. 22 Goiânia, Rio de Janeiro: Ministério das Minas e Energia. 768p. <https:...>

Renard, K. G., Foster, G. R., Weesies, G. A., McCool, D. K., Yoder, D. C. (1997). Predicting soil erosion by water: A guide to conservation planning with the Revised Universal Soil Loss Equation (RUSLE), Agriculture Handbook Number 703. <https:...>

USDA-Agricultural Research Service. (2013). Science Documentation Revised Universal Soil Loss Equation Version 2. <https://www.ars.usda.gov/ARSUserFiles/60600505/RUSLE/RUSLE2_Science_Doc.pdf>.

Wischmeier, W. and Smith, D. (1978). Predicting rainfall erosion losses: a guide to conservation planning. Agricultural Handbook No. 537. U.S. Department of Agriculture, Washington DC, USA. <https:...>
}

\keyword{Soil Conservation; Erosivity index; R-Factor; Universal Soil Loss Equation-USLE
}
